Separate types for (u)int and (u)int256 (#115)
* Front-end changes for separating (u)int and (u)int256

* Replaced (u)int256 as the unbound type with (u)int in code generation.

* Fixed signatures of built-in functions to type check with the correct types of integers.

* Removed always succeeding check that used deprecated API storageBytes.

* Made uint the expected type for various operations (was previously uint256):

- the argument of a dynamic array allocation
- the index of an array index access
- the index of a fixed bytes index access
- the index of a TypeType

* Allowed for unbound types to not produce a warning when used in exponentiation or left shift.

* Added code generation for revert blocks in auto-generated consructors.

* Updated commandline tests.

* Updated auction example.

* Replaced (u)int256 with (u)int in the SolidityABIJSON c++ test suite.

* Added a sorted list of command-line tests to be skipped.

These are not Solidity code, but have been picked up incorrectly by the extraction script.

*  Replaced (u)int256 with (u)int in the SolidityASTJSON c++ test suite.

* Replaced (u)int256 with (u)int in the SolidityOptimizer c++ test suite.

* Replaced (u)int256 with (u)int in the SolidityEndToEndTest c++ test suite.

Added some failing tests to the excluded list. These tests were failing before but their hash changed.

* Removed some unneeded over/underflow checks from SMTChecker and updated two related tests.

* Fixes in two SolidityEndToEnd tests.

Added a failing test to the excluded list due to hash change.

* Replaced (u)int256 with (u)int in the ABIDecoderTest c++ test suite.

* Fixed various tests in the SolidityNameAndTypeResolution c++ test suite.

Some new commandline tests have been added that all are supposed to not compile successfully, so their hashes have been added in the should-skip-tests list.

* Replaced (u)int256 with (u)int in the SolidityNatspecJSON c++ test suite.

* Added a few new passes after rebasing on top of the encodings and hashes.

* Fixed mapping c++ test that was failing due to treating uint256 as unbound.

* Clean up for the list of failing tests.

Removed a few hashes that came from extracted programs with code changed by this PR resulting in them having a new hash.

* Removed range check generation for when converting a uint256 to uint.

* Fix for static overflow check when using exponentiation or left shift on a literal.

* Made type of length member for fixed size arrays uint256.

* Added an assertion in the numBits API for IntegerType to ensure that it is never called for the unbound variants of the type.

Also added a few missing guards and replaced uses of private field m_bits with calls to numBits in places it made sense to have the assertion check fire.

+++ libsolidity/analysis/GlobalContext.cpp
@@ -35,7 +35,7 @@ namespace solidity
 
 GlobalContext::GlobalContext():
 m_magicVariables(vector<shared_ptr<MagicVariableDeclaration const>>{
-	make_shared<MagicVariableDeclaration>("addmod", make_shared<FunctionType>(strings{"uint256", "uint256", "uint256"}, strings{"uint256"}, FunctionType::Kind::AddMod, false, StateMutability::Pure)),
+	make_shared<MagicVariableDeclaration>("addmod", make_shared<FunctionType>(strings{"uint", "uint", "uint"}, strings{"uint"}, FunctionType::Kind::AddMod, false, StateMutability::Pure)),
 	make_shared<MagicVariableDeclaration>("assert", make_shared<FunctionType>(strings{"bool"}, strings{}, FunctionType::Kind::Assert, false, StateMutability::Pure)),
 	make_shared<MagicVariableDeclaration>("block", make_shared<MagicType>(MagicType::Kind::Block)),
 	make_shared<MagicVariableDeclaration>("ecrecover", make_shared<FunctionType>(strings{"bytes32", "uint8", "bytes32", "bytes32"}, strings{"address"}, FunctionType::Kind::ECRecover, false, StateMutability::Pure)),
@@ -47,7 +47,7 @@ m_magicVariables(vector<shared_ptr<MagicVariableDeclaration const>>{
 	make_shared<MagicVariableDeclaration>("log3", make_shared<FunctionType>(strings{"bytes32", "bytes32", "bytes32", "bytes32"}, strings{}, FunctionType::Kind::Log3)),
 	make_shared<MagicVariableDeclaration>("log4", make_shared<FunctionType>(strings{"bytes32", "bytes32", "bytes32", "bytes32", "bytes32"}, strings{}, FunctionType::Kind::Log4)),
 	make_shared<MagicVariableDeclaration>("msg", make_shared<MagicType>(MagicType::Kind::Message)),
-	make_shared<MagicVariableDeclaration>("mulmod", make_shared<FunctionType>(strings{"uint256", "uint256", "uint256"}, strings{"uint256"}, FunctionType::Kind::MulMod, false, StateMutability::Pure)),
+	make_shared<MagicVariableDeclaration>("mulmod", make_shared<FunctionType>(strings{"uint", "uint", "uint"}, strings{"uint"}, FunctionType::Kind::MulMod, false, StateMutability::Pure)),
 	make_shared<MagicVariableDeclaration>("now", make_shared<IntegerType>(256)),
 	make_shared<MagicVariableDeclaration>("require", make_shared<FunctionType>(strings{"bool"}, strings{}, FunctionType::Kind::Require, false, StateMutability::Pure)),
 	make_shared<MagicVariableDeclaration>("revert", make_shared<FunctionType>(strings(), strings(), FunctionType::Kind::Revert, false, StateMutability::Pure)),
